+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- issue when using `ToJson` where array of `ref<T>` was converted to 
+  `[ [ {}, ...] ]` instead of `[ {}, ... ]`.
 
 ------------------------